+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Fix `ConcurrentModificationException`s in store subscriptions' iteration by adding safe iteration over them.
+- Rename `toggleAbility` to `toggleEnabled` as the name was confusing.
 
 ## [1.1.2] - 2020-02-07
 ### Added
@@ -66,7 +67,7 @@ state (`success` or `failure`)
 ### Added
 - Initial architecture release.
 
-[Unreleased]: https://github.com/bq/mini-kotlin/compare/1.1.1...HEAD
+[Unreleased]: https://github.com/bq/mini-kotlin/compare/1.1.2...HEAD
 [1.1.2]: https://github.com/bq/mini-kotlin/compare/1.1.1...1.1.2
 [1.1.1]: https://github.com/bq/mini-kotlin/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/bq/mini-kotlin/compare/1.0.9...1.1.0